requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0